Move maps to their own component folders
