+++ src/app/src/components/AnimatedMap/AnimatedMapLegend.tsx
