+++ src/app/src/components/AnimatedMap/TimeControlIcon.tsx
