+++ src/app/src/components/AnimatedMap/dummySpendingDataByMonth.ts
