+++ src/app/src/components/AnimatedMap/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './AnimatedMap';
