+++ src/app/src/components/PerCapitaMap/PersonIcon.tsx
