+++ src/app/src/components/PerCapitaMap/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './PerCapitaMap';
